+++ API/RateSystem/src/main/java/com/java/RateSystem/controller/ServiceController.java
@@ -37,7 +37,7 @@ public class ServiceController {
     }
 
     @GetMapping("/{id}")
-    ResponseEntity<ResponseObject> findById(@PathVariable long id){
+    ResponseEntity<ResponseObject> findById(@PathVariable Integer id){
         Optional<Servicerate> foundProduct = productService.findById(id);
         return foundProduct.isPresent() ?
                 ResponseEntity.status(HttpStatus.OK).body(
@@ -106,7 +106,7 @@ public class ServiceController {
 
     //Update data
     @PutMapping("/{id}")
-    ResponseEntity<ResponseObject> updateService(@RequestBody Servicerate newService, @PathVariable long id){
+    ResponseEntity<ResponseObject> updateService(@RequestBody Servicerate newService, @PathVariable Integer id){
         Servicerate updateService =  productService.findById(id)
                 .map(service -> {
                     service.setName(newService.getName());
@@ -123,7 +123,7 @@ public class ServiceController {
     }
 
     @DeleteMapping("/{id}")
-    ResponseEntity<ResponseObject> deleteProduct (@PathVariable long id){
+    ResponseEntity<ResponseObject> deleteProduct (@PathVariable Integer id){
         Optional<Servicerate> foundService = productService.findById(id);
         if(foundService.isPresent())
         {
